+++ src/index.js
@@ -1,5 +1,6 @@
 import { Point } from './utils';
 import EventEmitter from '../../EventEmitter/index';
+import baseModule from '../../tatty-screen-base-module/index';
 
 export default class Screen extends EventEmitter {
 
@@ -13,17 +14,17 @@ export default class Screen extends EventEmitter {
      */
     constructor( el, opts, modules ) {
 
-        this.opts = Object.assign({
+        this.defaults = {
             cols: 80,
-            rows: 24,
-            scan: true,
-            scanOffset: 3
-        }, opts || {} );
+            rows: 24
+        };
 
         // Register modules early
         if ( Array.isArray( modules ) ) {
             this.registerModules( modules );
         }
+
+        this.opts = Object.assign( this.defaults, opts || {} );
 
         // Set DOM styles
         this.parent = el;
@@ -46,10 +47,6 @@ export default class Screen extends EventEmitter {
             y: -1
         });
         this.flashCursor();
-
-        if ( this.opts.scan ) {
-            this.scan = this.createScan();
-        }
 
         /**
          * Events
@@ -473,16 +470,6 @@ export default class Screen extends EventEmitter {
                 left: 0;
                 white-space: pre;
             }
-            .tatty .scan {
-                position: absolute;
-                left: 0;
-                top: 0;
-                right: 0;
-                bottom: 0;
-                pointer-events: none;
-                background-repeat: repeat;
-                opacity: .3;
-            }
         `;
 
         var head = document.querySelector( 'head' );
@@ -619,37 +606,14 @@ export default class Screen extends EventEmitter {
         return fontWidth;
     }
 
-
-    /**
-     * Create scanlines
-     */
-    createScan() {
-        var canvas = document.createElement( 'canvas' );
-        var scan = document.createElement( 'div' );
-        scan.classList.add( 'scan' );
-        this.parent.appendChild( scan );
-
-        var style = window.getComputedStyle( scan, null );
-
-        canvas.width = 1;
-        canvas.height = this.opts.scanOffset;
-        var ctx = canvas.getContext( '2d' );
-
-        ctx.fillStyle = style.color || '#fff';
-        ctx.fillRect( 0, this.opts.scanOffset - 1, 1, 1 );
-
-        scan.style.backgroundImage = 'url( ' + canvas.toDataURL() + ' )';
-
-        return scan;
-    }
-
     /**
      * Registers and mixes in screen modules
      */
     registerModules( modules ) {
         modules.forEach( function( module ) {
-            if ( typeof module !== 'tattyScreenModule' ) {
-                console.log( 'Error trying to attach module to tatty-screen' );
+            if ( !module instanceof baseModule ) {
+                console.log( 'Error trying to attach module to tatty-screen', module.name );
+                return;
             }
 
             if ( module.init ) {